+++ Models/Teacher.cs
@@ -29,7 +29,7 @@ namespace SHikkhanobishAPI.Models
         public int Age { get; set; }
         public string Class { get; set; }
         public string InstitutionName { get; set; }
-        public int RechargedAmount { get; set; }
+        public float RechargedAmount { get; set; }
         public int isFounder { get; set; }
         public string response { get; set; }
     }
